added logic for updating model, refactored model held in ram; TODO fix bug where added constraints are not properly returned from db-service to trace-receiver

+++ java/data-layer/src/main/java/at/scch/freiseisen/ma/data_layer/dto/ProbDeclareConstraint.java
@@ -1,4 +1,4 @@
 package at.scch.freiseisen.ma.data_layer.dto;
 
-public record ProbDeclareConstraint(float probability, String declareTemplate) {
+public record ProbDeclareConstraint(double probability, String declareTemplate) {
 }
